Better versioning during releasing from GHA
- Hard code the version number into src/odc/loader/__init__.py
- Check during a release that the hard coded version matches the git tag
- Push a commit immediately following the release which increments the
  version number and appends a dev tag to it.

This will require setting the version number in the source code *before*
creating a Release using GitHub. Possibly something to be addressed and
simplified in the future. It would be possible to set the verison number
into the code based on the release, but, I'll have to think about it.